requirements.txt
selenium
beautifulsoup4
requests
python-dotenv
deep-translator
langdetect
webdriver-manager